backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
openai>=1.12.0
httpx>=0.26.0
cachetools>=5.3.2
slowapi>=0.1.9
sentry-sdk[fastapi]>=1.40.0
PyJWT[crypto]>=2.8.0
